stratum: fix Win64 cross build — compat.h closesocket macro clash (#8)
MinGW cross failed: compat.h maps closesocket() to
myclosesocket(SOCKET&), which cannot bind the rvalue in Stop().
Use ::shutdown() on both platforms instead — it was already the
POSIX path, better matches the intent (unblock the blocked reader;
asio keeps ownership of the close), and dodges the macro entirely.
Widen the stored handle to intptr_t while here: Winsock SOCKET is
pointer-sized.

Slack generated by MWOWM

+++ src/stratum.h
@@ -117,8 +117,9 @@ private:
     boost::thread* pthreadClient;
     std::vector<boost::thread*> vWorkers;
     volatile bool fShutdown;
-    //! Raw fd of the live socket so Stop() can unblock a blocking read.
-    volatile int nSocketFd;
+    //! Raw handle of the live socket so Stop() can unblock a blocking read
+    //! (intptr_t: Winsock SOCKET is pointer-sized, not int).
+    volatile intptr_t nSocketFd;
 };
 
 /** Global client instance driven by -stratum=<host:port> / -stratumuser=<Q-address>
